coinbase_analytics/requirements.txt
beavers
black
confluent_kafka
isort
perspective-python
tabulate
tornado
pyarrow
websockets
